+++ Editor/Views/MonoGraphModelEditor.cs
@@ -1,0 +1,11 @@
+using System;
+using UnityEditor;
+using UnityEngine.UIElements;
+
+namespace NewGraph {
+    [CustomEditor(typeof(MonoGraphModel))]
+    public class MonoGraphModelEditor : GraphModelEditorBase {
+        protected override void CreateCustomizedListView(VisualElement inspector) {}
+    }
+}
+